Factored out ManagedConfigSaver
