+++ collect_app/src/main/java/org/odk/collect/android/mdm/ManagedConfigManager.kt
@@ -5,15 +5,8 @@ import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
 import android.content.RestrictionsManager
-import android.os.Bundle
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
-import org.odk.collect.android.projects.ProjectCreator
-import org.odk.collect.android.projects.SettingsConnectionMatcher
-import org.odk.collect.projects.ProjectsRepository
-import org.odk.collect.settings.ODKAppSettingsImporter
-import org.odk.collect.settings.SettingsProvider
-import org.odk.collect.settings.keys.MetaKeys.KEY_INSTALL_ID
 
 /**
  * Manages configuration changes from a mobile device management system.
@@ -21,23 +14,20 @@ import org.odk.collect.settings.keys.MetaKeys.KEY_INSTALL_ID
  * See android.content.APP_RESTRICTIONS in AndroidManifest for supported configuration keys.
  */
 class ManagedConfigManager(
-    private val settingsProvider: SettingsProvider,
-    private val projectsRepository: ProjectsRepository,
-    private val projectCreator: ProjectCreator,
-    private val settingsImporter: ODKAppSettingsImporter,
+    private val managedConfigSaver: ManagedConfigSaver,
     private val restrictionsManager: RestrictionsManager,
     private val context: Context
 ) : DefaultLifecycleObserver {
 
     private val restrictionsReceiver = object : BroadcastReceiver() {
         override fun onReceive(context: Context, intent: Intent) {
-            applyConfig()
+            managedConfigSaver.applyConfig(restrictionsManager.applicationRestrictions)
         }
     }
 
     override fun onResume(owner: LifecycleOwner) {
         super.onResume(owner)
-        applyConfig()
+        managedConfigSaver.applyConfig(restrictionsManager.applicationRestrictions)
 
         val restrictionsFilter = IntentFilter(Intent.ACTION_APPLICATION_RESTRICTIONS_CHANGED)
         context.registerReceiver(restrictionsReceiver, restrictionsFilter)
@@ -46,34 +36,5 @@ class ManagedConfigManager(
     override fun onPause(owner: LifecycleOwner) {
         super.onPause(owner)
         context.unregisterReceiver(restrictionsReceiver)
-    }
-
-    private fun applyConfig() {
-        val managedConfig = restrictionsManager.applicationRestrictions
-        applyDeviceId(managedConfig)
-        applySettingsJson(managedConfig)
-    }
-
-    private fun applyDeviceId(managedConfig: Bundle) {
-        if (managedConfig.containsKey(DEVICE_ID_KEY) && !managedConfig.getString(DEVICE_ID_KEY).isNullOrBlank()) {
-            settingsProvider.getMetaSettings().save(KEY_INSTALL_ID, managedConfig.getString(DEVICE_ID_KEY))
-        }
-    }
-
-    private fun applySettingsJson(managedConfig: Bundle) {
-        if (managedConfig.containsKey(SETTINGS_JSON_KEY) && !managedConfig.getString(SETTINGS_JSON_KEY).isNullOrBlank()) {
-            val settingsJson = managedConfig.getString(SETTINGS_JSON_KEY)
-
-            val settingsConnectionMatcher = SettingsConnectionMatcher(projectsRepository, settingsProvider)
-            when (val matchingProjectUUID = settingsJson?.let { settingsConnectionMatcher.getProjectWithMatchingConnection(it) }) {
-                null -> projectCreator.createNewProject(settingsJson!!)
-                else -> settingsImporter.fromJSON(settingsJson, projectsRepository.get(matchingProjectUUID)!!)
-            }
-        }
-    }
-
-    companion object {
-        private const val DEVICE_ID_KEY = "device_id"
-        private const val SETTINGS_JSON_KEY = "settings_json"
     }
 }
